Move get_class_members()
