+++ sphinx/ext/autodoc/_member_finder.py
@@ -1,0 +1,130 @@
+from __future__ import annotations
+
+from typing import TYPE_CHECKING
+
+from sphinx.errors import PycodeError
+from sphinx.ext.autodoc._sentinels import INSTANCE_ATTR, SLOTS_ATTR
+from sphinx.ext.autodoc.mock import ismock, undecorate
+from sphinx.pycode import ModuleAnalyzer
+from sphinx.util.inspect import (
+    getannotations,
+    getmro,
+    getslots,
+    isenumclass,
+    safe_getattr,
+)
+
+if TYPE_CHECKING:
+    from typing import Any, Protocol
+
+    from sphinx.ext.autodoc import ObjectMember
+
+    class _AttrGetter(Protocol):
+        def __call__(self, obj: Any, name: str, default: Any = ..., /) -> Any: ...
+
+
+def get_class_members(
+    subject: Any, objpath: Any, attrgetter: _AttrGetter, inherit_docstrings: bool = True
+) -> dict[str, ObjectMember]:
+    """Get members and attributes of target class."""
+    from sphinx.ext.autodoc._documenters import ObjectMember
+
+    # the members directly defined in the class
+    obj_dict = attrgetter(subject, '__dict__', {})
+
+    members: dict[str, ObjectMember] = {}
+
+    # enum members
+    if isenumclass(subject):
+        for name, defining_class, value in _filter_enum_dict(
+            subject, attrgetter, obj_dict
+        ):
+            # the order of occurrence of *name* matches the subject's MRO,
+            # allowing inherited attributes to be shadowed correctly
+            if unmangled := unmangle(defining_class, name):
+                members[unmangled] = ObjectMember(
+                    unmangled, value, class_=defining_class
+                )
+
+    # members in __slots__
+    try:
+        subject___slots__ = getslots(subject)
+        if subject___slots__:
+            for name, docstring in subject___slots__.items():
+                members[name] = ObjectMember(
+                    name, SLOTS_ATTR, class_=subject, docstring=docstring
+                )
+    except (TypeError, ValueError):
+        pass
+
+    # other members
+    for name in dir(subject):
+        try:
+            value = attrgetter(subject, name)
+            if ismock(value):
+                value = undecorate(value)
+
+            unmangled = unmangle(subject, name)
+            if unmangled and unmangled not in members:
+                if name in obj_dict:
+                    members[unmangled] = ObjectMember(unmangled, value, class_=subject)
+                else:
+                    members[unmangled] = ObjectMember(unmangled, value)
+        except AttributeError:
+            continue
+
+    try:
+        for cls in getmro(subject):
+            try:
+                modname = safe_getattr(cls, '__module__')
+                qualname = safe_getattr(cls, '__qualname__')
+                analyzer = ModuleAnalyzer.for_module(modname)
+                analyzer.analyze()
+            except AttributeError:
+                qualname = None
+                analyzer = None
+            except PycodeError:
+                analyzer = None
+
+            # annotation only member (ex. attr: int)
+            for name in getannotations(cls):
+                unmangled = unmangle(cls, name)
+                if unmangled and unmangled not in members:
+                    if analyzer and (qualname, unmangled) in analyzer.attr_docs:
+                        docstring = '\n'.join(analyzer.attr_docs[qualname, unmangled])
+                    else:
+                        docstring = None
+
+                    members[unmangled] = ObjectMember(
+                        unmangled, INSTANCE_ATTR, class_=cls, docstring=docstring
+                    )
+
+            # append or complete instance attributes (cf. self.attr1) if analyzer knows
+            if analyzer:
+                for (ns, name), docstring in analyzer.attr_docs.items():
+                    if ns == qualname and name not in members:
+                        # otherwise unknown instance attribute
+                        members[name] = ObjectMember(
+                            name,
+                            INSTANCE_ATTR,
+                            class_=cls,
+                            docstring='\n'.join(docstring),
+                        )
+                    elif (
+                        ns == qualname
+                        and docstring
+                        and isinstance(members[name], ObjectMember)
+                        and not members[name].docstring
+                    ):
+                        if cls != subject and not inherit_docstrings:
+                            # If we are in the MRO of the class and not the class itself,
+                            # and we do not want to inherit docstrings, then skip setting
+                            # the docstring below
+                            continue
+                        # attribute is already known, because dir(subject) enumerates it.
+                        # But it has no docstring yet
+                        members[name].docstring = '\n'.join(docstring)
+    except AttributeError:
+        pass
+
+    return members
