+++ sphinx/ext/autodoc/importer.py
@@ -24,7 +24,6 @@ from sphinx.ext.autodoc._property_types import (
     _ModuleProperties,
 )
 from sphinx.ext.autodoc._sentinels import (
-    INSTANCE_ATTR,
     RUNTIME_INSTANCE_ATTRIBUTE,
     SLOTS_ATTR,
     UNINITIALIZED_ATTR,
@@ -34,11 +33,7 @@ from sphinx.locale import __
 from sphinx.pycode import ModuleAnalyzer
 from sphinx.util import inspect, logging
 from sphinx.util.inspect import (
-    getannotations,
-    getmro,
-    getslots,
     isclass,
-    isenumclass,
     safe_getattr,
     unwrap_all,
 )
@@ -51,7 +46,6 @@ if TYPE_CHECKING:
     from typing import Any, Protocol
 
     from sphinx.environment import BuildEnvironment, _CurrentDocument
-    from sphinx.ext.autodoc import ObjectMember
     from sphinx.ext.autodoc._property_types import _AutodocFuncProperty, _AutodocObjType
 
     class _AttrGetter(Protocol):
@@ -403,113 +397,6 @@ def _import_from_module_and_path(
         errmsg = '; '.join(err_parts)
         logger.debug(errmsg)
         raise ImportError(errmsg) from exc
-
-
-def get_class_members(
-    subject: Any, objpath: Any, attrgetter: _AttrGetter, inherit_docstrings: bool = True
-) -> dict[str, ObjectMember]:
-    """Get members and attributes of target class."""
-    from sphinx.ext.autodoc._documenters import ObjectMember
-
-    # the members directly defined in the class
-    obj_dict = attrgetter(subject, '__dict__', {})
-
-    members: dict[str, ObjectMember] = {}
-
-    # enum members
-    if isenumclass(subject):
-        for name, defining_class, value in _filter_enum_dict(
-            subject, attrgetter, obj_dict
-        ):
-            # the order of occurrence of *name* matches the subject's MRO,
-            # allowing inherited attributes to be shadowed correctly
-            if unmangled := unmangle(defining_class, name):
-                members[unmangled] = ObjectMember(
-                    unmangled, value, class_=defining_class
-                )
-
-    # members in __slots__
-    try:
-        subject___slots__ = getslots(subject)
-        if subject___slots__:
-            for name, docstring in subject___slots__.items():
-                members[name] = ObjectMember(
-                    name, SLOTS_ATTR, class_=subject, docstring=docstring
-                )
-    except (TypeError, ValueError):
-        pass
-
-    # other members
-    for name in dir(subject):
-        try:
-            value = attrgetter(subject, name)
-            if ismock(value):
-                value = undecorate(value)
-
-            unmangled = unmangle(subject, name)
-            if unmangled and unmangled not in members:
-                if name in obj_dict:
-                    members[unmangled] = ObjectMember(unmangled, value, class_=subject)
-                else:
-                    members[unmangled] = ObjectMember(unmangled, value)
-        except AttributeError:
-            continue
-
-    try:
-        for cls in getmro(subject):
-            try:
-                modname = safe_getattr(cls, '__module__')
-                qualname = safe_getattr(cls, '__qualname__')
-                analyzer = ModuleAnalyzer.for_module(modname)
-                analyzer.analyze()
-            except AttributeError:
-                qualname = None
-                analyzer = None
-            except PycodeError:
-                analyzer = None
-
-            # annotation only member (ex. attr: int)
-            for name in getannotations(cls):
-                unmangled = unmangle(cls, name)
-                if unmangled and unmangled not in members:
-                    if analyzer and (qualname, unmangled) in analyzer.attr_docs:
-                        docstring = '\n'.join(analyzer.attr_docs[qualname, unmangled])
-                    else:
-                        docstring = None
-
-                    members[unmangled] = ObjectMember(
-                        unmangled, INSTANCE_ATTR, class_=cls, docstring=docstring
-                    )
-
-            # append or complete instance attributes (cf. self.attr1) if analyzer knows
-            if analyzer:
-                for (ns, name), docstring in analyzer.attr_docs.items():
-                    if ns == qualname and name not in members:
-                        # otherwise unknown instance attribute
-                        members[name] = ObjectMember(
-                            name,
-                            INSTANCE_ATTR,
-                            class_=cls,
-                            docstring='\n'.join(docstring),
-                        )
-                    elif (
-                        ns == qualname
-                        and docstring
-                        and isinstance(members[name], ObjectMember)
-                        and not members[name].docstring
-                    ):
-                        if cls != subject and not inherit_docstrings:
-                            # If we are in the MRO of the class and not the class itself,
-                            # and we do not want to inherit docstrings, then skip setting
-                            # the docstring below
-                            continue
-                        # attribute is already known, because dir(subject) enumerates it.
-                        # But it has no docstring yet
-                        members[name].docstring = '\n'.join(docstring)
-    except AttributeError:
-        pass
-
-    return members
 
 
 def _import_object(
